backend/requirements.txt
django-cors-headers
graphene_django
django_rest_framework

